put proj_loss in base class
